+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.3.7 Under development
 
+- Enh #54: Add `setAttribute()` prefixed attribute APIs to `HasContainerCollection`, `HasLabelCollection`, `HasPrefixCollection`, and `HasSuffixCollection` and related tests (@terabytesoftw)
+
 ## 0.3.6 February 14, 2026
 
 - Enh #51: Add `HasLabelCollection` mixin for managing label tag and attributes (@terabytesoftw)